# api/openapi.yaml
openapi: 3.0.3
info: {title: Netlab API, version: 1.0.0}
servers: [{url: /api/v1}]
security: [{bearer: []}]
paths:
...
  /nodes/{id}/retirement:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: nodeRetirement
      responses:
        '200': {description: Node exit plan and progress, content: {application/json: {schema: {$ref: '#/components/schemas/NodeRetirement'}}}}
    post:
      operationId: retireNode
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
        '409': {description: Node has dependencies to resolve}
    delete:
      operationId: resumeNodeScheduling
      responses:
        '204': {description: Scheduling resumed}
...
components:
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string}}
...
  responses:
    Accepted:
      description: Accepted
      content: {application/json: {schema: {$ref: '#/components/schemas/Operation'}}}
  securitySchemes:
    bearer: {type: http, scheme: bearer}
  schemas:
...
    Resources:
      type: object
      required: [cpu, memoryMiB, diskGiB]
      properties:
        cpu: {type: integer, minimum: 1}
        memoryMiB: {type: integer, format: int64, minimum: 64}
        diskGiB: {type: integer, format: int64, minimum: 1}
    Hardware:
      type: object
      required: [firmware, machine, diskBus, nicModel]
      properties:
        firmware: {type: string, enum: [bios, uefi]}
        machine: {type: string}
        diskBus: {type: string, enum: [ide, sata, scsi, virtio]}
        diskController: {type: string}
        nicModel: {type: string, enum: [virtio, e1000, e1000e, rtl8139, vmxnet3]}
        cpuModel: {type: string}
        cpuTopology: {$ref: '#/components/schemas/CpuTopology'}
        numaNodes: {type: integer, minimum: 1, description: Guest NUMA nodes; vCPUs and memory are divided evenly. Omitted keeps one NUMA node.}
        numaPlacement: {type: string, enum: [preferred, strict], description: Native automatic host CPU and memory placement. Omitted lets the host scheduler choose.}
        secureBoot: {type: boolean}
        tpm: {type: boolean}
        guestAgent: {type: boolean}
        firmwareCode: {type: string}
        firmwareVars: {type: string}
    CpuTopology:
      type: object
      required: [sockets, threads]
      properties:
        sockets: {type: integer, minimum: 1}
        threads: {type: integer, minimum: 1}
    Volume:
      type: object
      required: [id, mountPath, sizeGiB]
      properties:
        id: {type: string}
        mountPath: {type: string}
        sizeGiB: {type: integer, format: int64}
        persistentVolumeId: {type: string, description: Registered persistent volume. Its size comes from the volume catalogue.}
...
    NodeVolume:
      type: object
      required: [id, kind, sizeGiB, storage]
      properties:
        id: {type: string}
        kind: {$ref: '#/components/schemas/TemplateKind'}
        sizeGiB: {type: integer, format: int64}
        storage: {$ref: '#/components/schemas/StorageInfo'}
    NodeRetirement:
      type: object
      required: [assets, blockers, draining]
      properties:
        assets: {type: array, items: {$ref: '#/components/schemas/StoragePoolAsset'}}
        blockers: {type: array, items: {type: string}}
        draining: {type: boolean}
        operation: {$ref: '#/components/schemas/Operation'}
...
    TemplateDisk:
      type: object
      required: [id, sizeGiB, bus, bootOrder]
      properties:
        id: {type: string}
        sizeGiB: {type: integer, format: int64, minimum: 1}
        bus: {type: string, enum: [ide, sata, scsi, virtio]}
        bootOrder: {type: integer, minimum: 1}
        controllerModel: {type: string}
        controllerIndex: {type: integer, minimum: 0}
        controllerUnit: {type: integer, minimum: 0}
...
    TemplateFormat:
      type: string
      enum: [oci, docker, qcow2, raw, vmdk, ova, ovf, iso]
    TemplateInitialization:
      type: string
      enum: [none, cloud-init, cloudbase-init]
      default: none
    TemplateState:
      type: string
      enum: [importing, ready, failed, deleting]
      x-enum-varnames: [TemplateStateImporting, TemplateStateReady, TemplateStateFailed, TemplateStateDeleting]
    Template:
      type: object
      required: [id, name, kind, os, version, source, resources]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {$ref: '#/components/schemas/TemplateKind'}
        os: {type: string}
        version: {type: integer}
        source: {type: string}
        format: {$ref: '#/components/schemas/TemplateFormat'}
        artifactNodeId: {type: string, readOnly: true}
        operationId: {type: string, readOnly: true}
        resources: {$ref: '#/components/schemas/Resources'}
        hardware: {$ref: '#/components/schemas/Hardware'}
        disks: {type: array, items: {$ref: '#/components/schemas/TemplateDisk'}}
        media: {type: array, items: {$ref: '#/components/schemas/TemplateMedia'}}
        stateFiles: {type: array, items: {type: string}, readOnly: true}
        nicModels: {type: array, items: {type: string}}
        initialization: {$ref: '#/components/schemas/TemplateInitialization'}
        volumes: {type: array, items: {$ref: '#/components/schemas/Volume'}}
        state: {$ref: '#/components/schemas/TemplateState'}
        error: {type: string}
    TemplateMedia:
      type: object
      required: [id, source]
      properties:
        id: {type: string}
        source: {type: string}
...
    Interface:
      type: object
      required: [id, networkId, mac, address, primary]
      properties:
        id: {type: string}
        networkId: {type: string}
        mac: {type: string}
        address: {type: string}
        primary: {type: boolean}
    Asset:
      type: object
      required: [id, name, templateId, resources, interfaces]
      properties:
        id: {type: string}
        name: {type: string}
        templateId: {type: string}
        storagePoolId: {type: string, description: Omitted selects storage automatically; an existing asset keeps its storage until migrated.}
        resources: {$ref: '#/components/schemas/Resources'}
        interfaces: {type: array, items: {$ref: '#/components/schemas/Interface'}}
        volumes: {type: array, items: {$ref: '#/components/schemas/Volume'}}
        parameters: {type: object, additionalProperties: {type: string}}
        restartPolicy: {$ref: '#/components/schemas/RestartPolicy'}
        guest: {$ref: '#/components/schemas/GuestSettings'}
        media: {type: array, items: {type: string}, description: Attached template media IDs. Omitted attaches all; an empty array ejects all.}
        pciBinding: {$ref: '#/components/schemas/PciBinding'}
    PciBinding:
      type: object
      required: [nodeId, groupIds]
      properties:
        nodeId: {type: string}
        groupIds: {type: array, minItems: 1, uniqueItems: true, items: {type: string}}
...
    RestartPolicy:
      type: string
      enum: [never, on-failure, always]
      default: never
      description: Restart unexpected exits only. Explicit stop remains stopped, including across node restarts.
...
    GuestSettings:
      type: object
      properties:
        hostname: {type: string}
        username: {type: string}
        sshAuthorizedKeys: {type: array, items: {type: string}}
...
    TemplateKind:
      type: string
      enum: [container, vm]
...
    Operation:
      type: object
      required: [id, kind, state, phase, completed, total, retryable, createdAt, updatedAt]
      properties:
        id: {type: string}
        environmentId: {type: string}
        kind: {type: string}
        state: {type: string, enum: [queued, running, succeeded, failed, partially_applied]}
        phase: {type: string}
        completed: {type: integer}
        total: {type: integer}
        retryable: {type: boolean}
        error: {type: string}
        results: {type: array, items: {$ref: '#/components/schemas/ExecutionResult'}}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
...
    ResolvedInterface:
      type: object
      required: [id, networkId, mac, address, prefix, primary, portName, mtu]
      properties:
        id: {type: string}
        networkId: {type: string}
        mac: {type: string}
        address: {type: string}
        prefix: {type: integer}
        primary: {type: boolean}
        portName: {type: string}
        mtu: {type: integer}
        gateway: {type: string}
        dns: {type: array, items: {type: string}}
    AssetExecution:
      type: object
      required: [asset, template, instanceId, interfaces]
      properties:
        asset: {$ref: '#/components/schemas/Asset'}
        template: {$ref: '#/components/schemas/Template'}
        instanceId: {type: string}
        previousInstanceId: {type: string}
        dataSetId: {type: string, x-go-type-skip-optional-pointer: true, description: Actual writable data set; retained across ordinary rebuilds}
        volumeSources: {type: object, additionalProperties: {$ref: '#/components/schemas/NodeVolume'}, x-go-type-skip-optional-pointer: true}
        pciDevices: {type: array, items: {type: string}, x-go-type-skip-optional-pointer: true}
        storagePath: {type: string, description: Node-local managed storage selected by the scheduler}
        storagePoolId: {type: string, description: Actual selected pool; omitted for node default storage}
        storageFilesystem: {type: string, description: Physical filesystem reservation identity}
        rbd: {$ref: '#/components/schemas/RbdStorage'}
        interfaces: {type: array, items: {$ref: '#/components/schemas/ResolvedInterface'}}
...
    ExecutionResult:
      type: object
      required: [assetId, instanceId, state, observedAt]
      properties:
        environmentId: {type: string}
        nodeId: {type: string}
        assetId: {type: string}
        instanceId: {type: string}
        state: {type: string}
        error: {type: string}
        execution: {$ref: '#/components/schemas/AssetExecution'}
        recovery: {$ref: '#/components/schemas/RecoveryCapture'}
        observedAt: {type: string, format: date-time}
    RecoveryConsistency:
      type: string
      enum: [crash, filesystem, application]
    RecoveryCapture:
      type: object
      required: [sizeBytes, memory, consistency]
      properties:
        sizeBytes: {type: integer, format: int64}
        memory: {type: boolean}
        consistency: {$ref: '#/components/schemas/RecoveryConsistency'}
...
    StorageInfo:
      type: object
      required: [path, filesystem, capacityBytes, availableBytes]
      properties:
        path: {type: string}
        filesystem: {type: string}
        capacityBytes: {type: integer, format: int64}
        availableBytes: {type: integer, format: int64}
        nativeSnapshots: {type: boolean, x-go-type-skip-optional-pointer: true}
        rbd: {$ref: '#/components/schemas/RbdStorage'}
...
    StoragePoolAsset:
      type: object
      required: [environmentId, environmentName, revision, assetId, assetName, nodeId, sizeGiB, state]
      properties:
        environmentId: {type: string}
        environmentName: {type: string}
        revision: {type: integer, format: int64}
        assetId: {type: string}
        assetName: {type: string}
        nodeId: {type: string}
        sizeGiB: {type: integer, format: int64}
        state: {type: string}
    RbdStorage:
      type: object
      required: [pool, imagePrefix, monitors, user, fsid, secretId]
      properties:
        pool: {type: string}
        imagePrefix: {type: string}
        monitors: {type: array, items: {type: string}}
        user: {type: string}
        fsid: {type: string}
        secretId: {type: string}
...