# db/queries/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package queries

import (
	"net/netip"

	"github.com/jackc/pgx/v5/pgtype"
)

type Backup struct {
	ID            string
	EnvironmentID *string
	RepositoryID  string
	Name          string
	Definition    []byte
	Result        []byte
	SizeBytes     int64
	State         string
	OperationID   string
	CreatedAt     pgtype.Timestamptz
}

type BackupRepository struct {
	ID          string
	NodeID      string
	Name        string
	Location    string
	Credentials []byte
	NativeID    *string
	State       string
	OperationID *string
	CreatedAt   pgtype.Timestamptz
}

type Blueprint struct {
	ID        string
	ProjectID string
	OwnerID   *string
	Name      string
	Version   int32
	CreatedAt pgtype.Timestamptz
	UpdatedAt pgtype.Timestamptz
}

type BlueprintVersion struct {
	ID                  string
	BlueprintID         string
	Version             int32
	Spec                []byte
	View                []byte
	AssetCount          int32
	NetworkCount        int32
	SourceEnvironmentID string
	SourceRevision      int32
	CreatedAt           pgtype.Timestamptz
}

type CaptureSegment struct {
	CaptureID     string
	EnvironmentID string
	NodeID        string
	CreatedAt     pgtype.Timestamptz
}

type Credential struct {
	Hash        []byte
	PrincipalID string
	ExpiresAt   pgtype.Timestamptz
	CreatedAt   pgtype.Timestamptz
}

type Environment struct {
	ID                 string
	ProjectID          string
	OwnerID            *string
	Name               string
	ExternalReference  *string
	Revision           int32
	Status             string
	Spec               []byte
	AppliedSpec        []byte
	View               []byte
	Draft              []byte
	NetworkNodeID      *string
	OperationID        *string
	Error              *string
	ClientRequestID    *string
	CreatedAt          pgtype.Timestamptz
	UpdatedAt          pgtype.Timestamptz
	BlueprintVersionID *string
	GatewayAddress     *netip.Addr
	VpnPublicKey       *string
	VpnMtu             *int32
}

type Event struct {
	Cursor        int64
	EnvironmentID *string
	Kind          string
	Payload       []byte
	CreatedAt     pgtype.Timestamptz
}

type ExternalNetworkLease struct {
	NodeID        string
	Interface     string
	Vlan          int32
	EnvironmentID string
	NetworkID     string
}

type Grant struct {
	PrincipalID string
	ScopeKind   string
	ScopeID     string
	Permissions []string
}

type GuestConnection struct {
	PrincipalID   string
	EnvironmentID string
	AssetID       string
	Encrypted     []byte
	Protocol      string
}

type MigrationAssetReservation struct {
	EnvironmentID string
	NodeID        interface{}
	Execution     interface{}
	Cpu           int32
	MemoryMib     int64
	DiskGib       int64
	VolumeGib     int64
}

type Node struct {
	ID               string
	Name             string
	Endpoint         string
	Info             []byte
	CapacityOverride []byte
	State            string
	ObservedAt       pgtype.Timestamptz
	Retiring         bool
}

type NodeAssetReservation struct {
	NodeID    string
	Cpu       int32
	MemoryMib int64
	DiskGib   int64
}

type Operation struct {
	ID               string
	EnvironmentID    *string
	ScopeKind        string
	ScopeID          string
	Kind             string
	AssetID          *string
	State            string
	Phase            string
	Payload          []byte
	Results          []byte
	Error            *string
	ExpectedRevision int32
	LeaseOwner       *string
	LeaseUntil       pgtype.Timestamptz
	CreatedAt        pgtype.Timestamptz
	UpdatedAt        pgtype.Timestamptz
	ClientRequestID  *string
}

type PersistentVolume struct {
	ID            string
	NodeID        string
	StoragePoolID string
	Name          string
	Kind          string
	SizeGib       int64
	State         string
	OperationID   *string
}

type Principal struct {
	ID            string
	Name          string
	Kind          string
	PasswordHash  []byte
	Administrator bool
	Disabled      bool
	CreatedAt     pgtype.Timestamptz
}

type Project struct {
	ID   string
	Name string
}

type RecoveryPoint struct {
	ID            string
	EnvironmentID string
	Name          string
	Revision      int32
	State         string
	Definition    []byte
	AssetCount    int32
	SizeBytes     int64
	OperationID   string
	CreatedAt     pgtype.Timestamptz
}

type RuntimeAsset struct {
	EnvironmentID string
	AssetID       string
	InstanceID    string
	NodeID        string
	Execution     []byte
	State         string
	Cpu           int32
	MemoryMib     int64
	DiskGib       int64
	Current       bool
	Error         *string
	ObservedAt    pgtype.Timestamptz
}

type ServicePort struct {
	ID            int64
	NodeID        string
	Protocol      string
	Port          *int32
	EnvironmentID string
	ServiceID     string
	OperationID   string
	State         string
	UpdatedAt     pgtype.Timestamptz
	Purpose       string
}

type StoragePool struct {
	ID          string
	Name        string
	Directory   string
	Path        string
	State       string
	OperationID *string
	CreatedAt   pgtype.Timestamptz
	NodeIds     []string
	Driver      string
	Managed     bool
}

type Template struct {
	ID         string
	Definition []byte
	CreatedAt  pgtype.Timestamptz
}

type VpnAccess struct {
	ID            string
	EnvironmentID string
	Definition    []byte
	Addresses     []netip.Addr
	Applied       bool
	OperationID   string
	CreatedAt     pgtype.Timestamptz
}

type VpnAlias struct {
	EnvironmentID string
	NetworkID     string
	Prefix        netip.Prefix
}